modularized event form and drink selections
